requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.6.0
django-markdownx==3.0.1
djangorestframework==3.12.2
gunicorn==20.0.4
Markdown==3.3.3
Pillow==8.1.0
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0
zipp==3.4.0
psycopg2==2.7.6.1
